refactor(runner): split start job discovery
